backend/pyproject.toml
[project]
name = "fundraiser-backend"
version = "0.1.0"
description = "FastAPI backend for fundraiser platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]